Moved GlossTests from Example to GlossFramework
